+++ src/main/java/pl/pw/radeja/speex/encoders/HideF0EncoderFirstMiddleRand.java
@@ -5,7 +5,7 @@ import pl.pw.radeja.speex.pitch.changers.LinearApproximateMiddleRand;
 
 public class HideF0EncoderFirstMiddleRand extends HideF0EncoderFirstMiddle {
 
-    public HideF0EncoderFirstMiddleRand(int threshold, String path) {
+    public HideF0EncoderFirstMiddleRand(float threshold, String path) {
         super(new LinearApproximateMiddleRand(threshold, 1L), path);
         this.type = Config.HideF0Type.FIRST_MIDDLE_RAND;
     }
